Put fixtures in the tests init file
