+++ bastion/tests/utils/test_role_manager.py
@@ -1,36 +1,6 @@
 from bastion.utils.role_manager import RoleManager
 
-
-class MockRole():
-    _id = 500
-
-    def __init__(self, name):
-        self.name = name
-        self.id = MockRole._id
-        MockRole._id += 1
-
-_everyone = MockRole('@everyone')
-_mock_bot = MockRole('MockBot')
-_some_role = MockRole('some_role')
-_another_role = MockRole('another_role')
-
-class MockUser():
-    def __init__(self):
-        self.id = 100
-
-
-class MockBot():
-    def __init__(self):
-        self.id = 1
-        self.user = MockUser()
-        self.roles = [_everyone, _mock_bot]
-
-
-class MockGuild():
-    def __init__(self):
-        self.id = 1000
-        self.members = [MockBot()]
-        self.roles = [_everyone, _some_role, _another_role, _mock_bot, MockRole('after_role')]
+from bastion.tests import MockBot, MockGuild, _everyone, _another_role, _some_role, _mock_bot_role
 
 
 def test_register_bot_roles():
@@ -40,7 +10,7 @@ def test_register_bot_roles():
 
     manager.register_bot_roles(bot, guild)
 
-    assert manager.role_map[guild.id] == [_mock_bot.id]
+    assert manager.role_map[guild.id] == [_mock_bot_role.id]
 
 
 def test_get_available_roles():
@@ -58,7 +28,7 @@ def test_get_available_roles_may_return_empty():
     manager = RoleManager()
     bot = MockBot()
     guild = MockGuild()
-    guild.roles = [_everyone, _mock_bot, _some_role]
+    guild.roles = [_everyone, _mock_bot_role, _some_role]
     manager.register_bot_roles(bot, guild)
 
     available_roles = manager.get_available_roles(guild)
